+++ source/pkgi_config.c
@@ -112,8 +112,9 @@ void pkgi_load_config(Config* config, char* refresh_url, uint32_t refresh_len)
     config->sort = SortByName;
     config->order = SortAscending;
     config->filter = DbFilterAll;
-    config->no_version_check = 0;
+    config->version_check = 1;
     config->dl_mode_background = 0;
+    config->music = 1;
 
     char data[4096];
     char path[256];
@@ -173,11 +174,15 @@ void pkgi_load_config(Config* config, char* refresh_url, uint32_t refresh_len)
             }
             else if (pkgi_stricmp(key, "no_version_check") == 0)
             {
-                config->no_version_check = 1;
+                config->version_check = 0;
             }
             else if (pkgi_stricmp(key, "dl_mode_background") == 0)
             {
                 config->dl_mode_background = 1;
+            }
+            else if (pkgi_stricmp(key, "no_music") == 0)
+            {
+                config->music = 0;
             }
         }
     }
@@ -243,7 +248,7 @@ void pkgi_save_config(const Config* config, const char* update_url)
     }
     len += pkgi_snprintf(data + len, sizeof(data) - len, "\n");
 
-    if (config->no_version_check)
+    if (!config->version_check)
     {
         len += pkgi_snprintf(data + len, sizeof(data) - len, "no_version_check 1\n");
     }
@@ -251,6 +256,11 @@ void pkgi_save_config(const Config* config, const char* update_url)
     if (config->dl_mode_background)
     {
         len += pkgi_snprintf(data + len, sizeof(data) - len, "dl_mode_background 1\n");
+    }
+
+    if (!config->music)
+    {
+        len += pkgi_snprintf(data + len, sizeof(data) - len, "no_music 1\n");
     }
 
     char path[256];